Standardize CI and move unit tests to Unit folder
