+++ .gitignore
@@ -50,19 +50,14 @@ nbproject
 .well-known
 /.buildpath
 
-# Tests - auto-generated files
-/tests/coverage*
-/tests/clover.xml
-/tests/output/clover.xml
-/tests/js/node_modules
-/tests/output/
-
-build
-
+/build/
+/l10n/.transifexrc
 .php_cs.cache
 
 # Composer
-composer.phar
 vendor/
 vendor-bin/**/composer.lock
 vendor-bin/**/vendor
+
+# Tests - auto-generated files
+/tests/output
