+++ vendor-bin/owncloud-codestyle/composer.json
@@ -1,9 +1,4 @@
 {
-  "config" : {
-    "platform": {
-      "php": "5.6.37"
-    }
-  },
   "require": {
     "owncloud/coding-standard": "^1.0"
   }
